requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic==1.10.13
python-dotenv==1.0.0
mysql-connector-python==8.2.0
openai==1.3.5
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
